Hey guys I decided to write some preliminary code for our assignment.
I made a member class with all the attributes we talked about on Monday. I also made a struct called memberNode as well as functions that allow us to make and alter an ordered linked list of member objects

I'm pretty sure we will be able to just copy/paste a lot of this into QT when ever we become familiar with it. But I'm sure we will have to change some (or a lot) stuff depending on what QT syntax is like.